+++ src/ui/svelte/window_manager/index.ts
@@ -1,10 +1,10 @@
+import { mount } from "svelte";
 import { startThemingTool } from "@seelen-ui/lib";
 import { getRootContainer } from "@shared";
 import { declareDocumentAsLayeredHitbox } from "@shared/layered";
 import { disableAnimationsOnPerformanceMode } from "@shared/performance";
-import { createRoot } from "react-dom/client";
 
-import { App } from "./app.tsx";
+import App from "./App.svelte";
 
 import "@shared/styles/colors.css";
 import "@shared/styles/reset.css";
@@ -15,4 +15,6 @@ await startThemingTool();
 disableAnimationsOnPerformanceMode();
 
 const container = getRootContainer();
-createRoot(container).render(<App />);
+mount(App, {
+  target: container,
+});